Fix: rubix-plat-build directory
