+++ src/system/apps/__init__.py
@@ -1,3 +1,3 @@
-__all__ = ['bacnet_server_app', 'lora_raw_app', 'point_server_app', 'wires_builds_app']
+__all__ = ['bacnet_server_app', 'lora_raw_app', 'point_server_app', 'wires_builds_app', 'rubix_plat_build_app']
 
 from src.system.apps import *
